Test refactor.
