+++ driver/test/sk/csirt/viruschecker/driver/routing/RoutingTest.kt
@@ -1,27 +1,29 @@
-package sk.csirt.viruschecker.driver
+package sk.csirt.viruschecker.driver.routing
 
 import com.xenomachina.argparser.ArgParser
-import io.ktor.http.HttpMethod
-import io.ktor.http.HttpStatusCode
+import io.ktor.locations.KtorExperimentalLocationsAPI
 import io.ktor.server.testing.TestApplicationEngine
-import io.ktor.server.testing.handleRequest
 import io.ktor.server.testing.withTestApplication
+import io.ktor.util.KtorExperimentalAPI
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import sk.csirt.viruschecker.config.filterArgsForArgParser
 import sk.csirt.viruschecker.driver.config.AntivirusType
 import sk.csirt.viruschecker.driver.config.CommandLineArguments
-import sk.csirt.viruschecker.routing.DriverRoutes
 import sk.csirt.viruschecker.driver.config.testDriverDependencyInjectionModule
-import kotlin.test.Test
-import kotlin.test.assertEquals
+import sk.csirt.viruschecker.driver.module
+import sk.csirt.viruschecker.driver.parsedArgs
 
+@KtorExperimentalLocationsAPI
+@KtorExperimentalAPI
 @ExperimentalCoroutinesApi
-fun <R> testApplication(test: TestApplicationEngine.() -> R): R =
+fun <R> createTestApplication(test: TestApplicationEngine.() -> R): R =
     withTestApplication({ module(testDriverDependencyInjectionModule) }) { this.test() }
 
 
+@KtorExperimentalLocationsAPI
+@KtorExperimentalAPI
 @ExperimentalCoroutinesApi
-class ApplicationTest {
+abstract class RoutingTest {
     init {
         parsedArgs = ArgParser(
             filterArgsForArgParser(
@@ -30,14 +32,5 @@ class ApplicationTest {
                 )
             )
         ).parseInto(::CommandLineArguments)
-    }
-
-    @Test
-    fun testRoot() {
-        testApplication {
-            handleRequest(HttpMethod.Get, DriverRoutes.index).apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-            }
-        }
     }
 }
